+++ R/tinyplot.R
@@ -13,10 +13,13 @@
 #'   'Examples' section below, or the function
 #'   \code{\link[grDevices]{xy.coords}} for details. If supplied separately, `x`
 #'   and `y` must be of the same length.
-#' @param by grouping variable(s). By default, groups will be represented
-#'   through colouring of the plot elements. However, this can be turned off
-#'   and other plot parameters (e.g., line types) can also take on grouping
-#'   behaviour via the special "by" keyword. See Examples.
+#' @param by grouping variable(s). The default behaviour is for groups to be
+#'   represented in the form of distinct colours, which will also trigger an
+#'   automatic legend. (See `legend` below for customization options.) However,
+#'   groups can also be presented through other plot parameters (e.g., `pch` or
+#'   `lty`) by passing an appropriate "by" keyword; see Examples. Note that
+#'   continuous (i.e., gradient) colour legends are also supported if the user
+#'   passes a numeric or integer to `by`.
 #' @param facet the faceting variable(s) that you want arrange separate plot
 #'   windows by. Can be specified in various ways:
 #'   - In "atomic" form, e.g. `facet = fvar`. To facet by multiple variables in 
@@ -100,11 +103,16 @@
 #'   with better default behaviour. Default is not to draw a grid.
 #' @param asp the y/xy/x aspect ratio, see `plot.window`.
 #' @param palette one of the following options:
-#'    - NULL (default), in which case the palette will be determined by the
-#'    the user's default graphics palette, e.g. "R4". See `?palette()`. Note
-#'    that some internal checking is done to make sure that resulting colours
-#'    match the number of groups. For larger group numbers, the "viridis"
-#'    palette will be used instead.
+#'    - NULL (default), in which case the palette will be chosen according to
+#'    the class and cardinality of the "by" grouping variable. For non-ordered
+#'    factors or strings with a reasonable number of groups, this will inherit
+#'    directly from the user's default \code{\link[grDevices]{palette}} (e.g.,
+#'    "R4"). In other cases, including ordered factors and high cardinality, the
+#'    "Viridis" palette will be used instead. Note that a slightly restricted
+#'    version of the "Viridis" palette---where extreme color values have been
+#'    trimmed to improve visual perception---will be used for ordered factors
+#'    and continuous variables. In the latter case of a continuous grouping
+#'    variable, we also generate a gradient legend swatch.
 #'    - A convenience string corresponding to one of the many palettes listed by
 #'    either `palette.pals()` or `hcl.pals()`. Note that the string can be
 #'    case-insensitive (e.g., "Okabe-Ito" and "okabe-ito" are both valid).
@@ -198,7 +206,7 @@
 #' out existing `plot` calls for `tinyplot` (or its shorthand alias `plt`),
 #' without causing unexpected changes to the output.
 #'   
-#' @importFrom grDevices adjustcolor extendrange palette palette.colors palette.pals hcl.colors hcl.pals xy.coords
+#' @importFrom grDevices adjustcolor colorRampPalette extendrange palette palette.colors palette.pals hcl.colors hcl.pals xy.coords
 #' @importFrom graphics abline arrows axis Axis box grconvertX grconvertY lines par plot.default plot.new plot.window points polygon segments title mtext text rect
 #' @importFrom utils modifyList head tail
 #' @importFrom stats na.omit
@@ -232,8 +240,8 @@
 #' 
 #' # (Notice that we also get an automatic legend.)
 #' 
-#' # You can also use the equivalent shorthand `plt` alias if you'd like to save
-#' # on a few keystrokes
+#' # You can also use the equivalent shorthand `plt()` alias if you'd like to
+#' # save on a few keystrokes
 #' 
 #' plt(Temp ~ Day | Month, data = aq) ## shorthand alias
 #'
@@ -445,13 +453,13 @@ tinyplot.default = function(
     # type = "density" plots (to make consistent with regular plot)
     if (is.null(fargs$main)) fargs$main = NA
     ## Catch for atomic density type to avoid "by" as legend title
-    if (is.null(fargs[["legend.args"]][["title"]])) {
-      fargs[["legend.args"]][["title"]] = by_dep
+    if (is.null(fargs[["legend_args"]][["title"]])) {
+      fargs[["legend_args"]][["title"]] = by_dep
     }
     ## Another catch for bespoke legend position (if originally passed via the formula method)
-    if (!is.null(fargs[["legend"]]) && !is.null(fargs[["legend.args"]])) {
+    if (!is.null(fargs[["legend"]]) && !is.null(fargs[["legend_args"]])) {
       if (names(fargs[["legend"]])[1] == "") names(fargs[["legend"]])[1] = "x"
-      fargs[["legend.args"]] = utils::modifyList(fargs[["legend"]], fargs[["legend.args"]])
+      fargs[["legend_args"]] = utils::modifyList(fargs[["legend"]], fargs[["legend_args"]])
       fargs[["legend"]] = NULL
     }
     fargs$y = fargs$ymin = fargs$ymax = fargs$ylab = fargs$xlab = NULL
@@ -529,8 +537,17 @@ tinyplot.default = function(
   if (!is.null(ymin)) ylim[1] = min(c(ylim, ymin))
   if (!is.null(ymax)) ylim[2] = max(c(ylim, ymax))
 
-
-  if (!is.null(by)) {
+  by_ordered = FALSE
+  by_continuous = !is.null(by) && inherits(by, c("numeric", "integer"))
+  # manual overrides with warning
+  if (isTRUE(by_continuous) && type %in% c("l", "b", "o", "ribbon", "polygon")) {
+    warning("\nContinuous legends not supported for this plot type. Reverting to discrete legend.")
+    by_continuous = FALSE
+  } else if (!is.null(by)){
+    by_ordered = is.ordered(by)
+  }
+  
+  if (!is.null(by) && !by_continuous) {
     split_data = list(x=x, y=y)
     split_data[["ymin"]] = ymin
     split_data[["ymax"]] = ymax
@@ -551,16 +568,20 @@ tinyplot.default = function(
   col = by_col(
     ngrps = ngrps,
     col = col,
-    palette = substitute(palette)
+    palette = substitute(palette),
+    gradient = by_continuous,
+    ordered = by_ordered
   )
   if (is.null(bg) && !is.null(fill)) bg = fill
-  if (!is.null(bg) && bg == "by") {
+  if (!is.null(bg) && length(bg)==1 && bg == "by") {
     bg = by_col(
       ngrps = ngrps,
       col = NULL,
-      palette = substitute(palette)
+      palette = substitute(palette),
+      gradient = by_continuous,
+      ordered = by_ordered
     )
-  } else {
+  } else if (length(bg) != ngrps) {
     bg = rep(bg, ngrps)
   }
   if (type == "ribbon") {
@@ -569,6 +590,27 @@ tinyplot.default = function(
     } else if (!is.null(col)) {
       bg = adjustcolor(col, ribbon_alpha)
     }
+  }
+  
+  if (isTRUE(by_continuous)) {
+    ## Identify the pretty break points for our labels
+    nlabs = 5
+    ncolors = length(col)
+    ubyvar = unique(by)
+    byvar_range = range(ubyvar)
+    pbyvar = pretty(byvar_range, n = nlabs)
+    pbyvar = pbyvar[pbyvar >= byvar_range[1] & pbyvar <= byvar_range[2]]
+    # optional thinning
+    if (length(ubyvar)==2 && all(ubyvar %in% pbyvar)) {
+      pbyvar = ubyvar
+    } else if (length(pbyvar)>nlabs) {
+      pbyvar = pbyvar[seq_along(pbyvar) %% 2 == 0]
+    }
+    ## Find the (approximate) location of our pretty labels
+    pidx = rescale_num(c(byvar_range, pbyvar), to = c(1, ncolors))[-c(1:2)]
+    pidx = round(pidx)
+    lgnd_labs = rep(NA, times = ncolors)
+    lgnd_labs[pidx] = pbyvar
   }
   
   # Determine the number and arrangement of facets.
@@ -635,15 +677,15 @@ tinyplot.default = function(
   # place and draw the legend
   has_legend = FALSE # simple indicator variable for later use
   
-  if (!exists("legend.args")) {
-    legend.args = dots[["legend.args"]]
+  if (!exists("legend_args")) {
+    legend_args = dots[["legend_args"]]
   }
-  if (is.null(legend.args)) legend.args = list(x = NULL)
+  if (is.null(legend_args)) legend_args = list(x = NULL)
   legend = substitute(legend)
   
   if (isFALSE(legend)) {
     legend = "none"
-    legend.args[["x"]] = "none"
+    legend_args[["x"]] = "none"
   }
   if (isTRUE(legend)) {
     legend = NULL
@@ -652,28 +694,30 @@ tinyplot.default = function(
   if (is.null(by)) {
     if (is.null(legend)) {
       legend = "none"
-      legend.args[["x"]] = "none"
+      legend_args[["x"]] = "none"
     }
   }
   
   if ((is.null(legend) || legend != "none") && isFALSE(add)) {
     
-    if (ngrps>1) {
-      lgnd_labs = names(split_data)
-    } else {
-      lgnd_labs = ylab
+    if (isFALSE(by_continuous)) {
+      if (ngrps>1) {
+        lgnd_labs = names(split_data)
+      } else {
+        lgnd_labs = ylab
+      }
     }
     
     has_sub = !is.null(sub)
     
     if (isTRUE(was_area_type)) {
-      legend.args[["pt.lwd"]] = par("lwd")
-      legend.args[["lty"]] = 0
+      legend_args[["pt.lwd"]] = par("lwd")
+      legend_args[["lty"]] = 0
     }
     
     draw_legend(
       legend = legend,
-      legend.args = legend.args,
+      legend_args = legend_args,
       by_dep = by_dep,
       lgnd_labs = lgnd_labs,
       type = type,
@@ -681,13 +725,14 @@ tinyplot.default = function(
       lty = lty,
       col = col,
       bg = bg,
+      gradient = by_continuous,
       cex = cex * cex_fct_adj,
       has_sub = has_sub
-      )
+    )
     
     has_legend = TRUE
     
-  } else if (legend.args[["x"]]=="none" && isFALSE(add)) {
+  } else if (legend_args[["x"]]=="none" && isFALSE(add)) {
     
     omar = par("mar")
     ooma = par("oma")
@@ -729,7 +774,7 @@ tinyplot.default = function(
     if (is.null(legend_eval)) {
       legend_eval = tryCatch(paste0(legend)[[2]], error = function(e) NULL)
     }
-    adj_title = !is.null(legend) && (legend == "top!" || (!is.null(legend.args[["x"]]) && legend.args[["x"]]=="top!") || (is.list(legend_eval) && legend_eval[[1]]=="top!") )
+    adj_title = !is.null(legend) && (legend == "top!" || (!is.null(legend_args[["x"]]) && legend_args[["x"]]=="top!") || (is.list(legend_eval) && legend_eval[[1]]=="top!") )
     if (is.null(main) || isFALSE(adj_title)) {
       title(
         main = main,
@@ -869,7 +914,7 @@ tinyplot.default = function(
       }
       
       ## Set the plot window
-      ## Problem: Passing extra args through ... (e.g., legend.args) to plot.window
+      ## Problem: Passing extra args through ... (e.g., legend_args) to plot.window
       ## triggers an annoying warning about unrecognized graphical params.
       # plot.window(
       #   xlim = xlim, ylim = ylim, 
@@ -1044,15 +1089,36 @@ tinyplot.default = function(
       idata[["facet"]] = NULL ## Don't need this anymore since we'll be splitting by ifacet
       ## Need extra catch for non-groupby data that also doesn't have ymin or
       ## ymax vars
-      if (is.null(by)) {
+      if (is.null(by) || isTRUE(by_continuous)) {
         if (is.null(idata[["ymin"]])) idata[["ymin"]] = NULL
         if (is.null(idata[["ymax"]])) idata[["ymax"]] = NULL
+      }
+      if (isTRUE(by_continuous)) {
+        idata[["col"]] = col[round(rescale_num(by, to = c(1,100)))]
+        idata[["bg"]] = bg[round(rescale_num(by, to = c(1,100)))]
       }
       idata = lapply(idata, split, ifacet)
       idata = do.call(function(...) Map("list", ...), idata)
     } else {
       idata = list(idata)
+      if (isTRUE(by_continuous)) {
+        if (length(col)!=1) {
+          idata[[1]][["col"]] = col[round(rescale_num(by, to = c(1,100)))]
+        } else {
+          idata[[1]][["col"]] = col
+        }
+        if (length(bg)!=1) {
+          idata[[1]][["bg"]] = bg[round(rescale_num(by, to = c(1,100)))]
+        } else {
+          idata[[1]][["bg"]] = bg
+        }
+      }
     }
+    
+    icol = col[i]
+    ibg = bg[i]
+    ipch = pch[i]
+    ilty = lty[i]
     
     ## Inner loop over the "facet" variables
     for (ii in seq_along(idata)) {
@@ -1060,6 +1126,11 @@ tinyplot.default = function(
       yy = idata[[ii]]$y
       yymin = idata[[ii]]$ymin
       yymax = idata[[ii]]$ymax
+      
+      if (isTRUE(by_continuous)) {
+        icol = idata[[ii]]$col
+        ibg = idata[[ii]]$bg
+      }
       
       # Set the facet "window" manually
       # See: https://github.com/grantmcdermott/tinyplot/issues/65
@@ -1090,8 +1161,8 @@ tinyplot.default = function(
           y0 = yymin,
           x1 = xx,
           y1 = yymax,
-          col = col[i],
-          lty = lty[i]
+          col = icol,
+          lty = ilty
         )
       }
       if (type == "errorbar") {
@@ -1100,8 +1171,8 @@ tinyplot.default = function(
           y0 = yymin,
           x1 = xx,
           y1 = yymax,
-          col = col[i],
-          lty = lty[i],
+          col = icol,
+          lty = ilty,
           length = 0.05,
           angle = 90,
           code = 3
@@ -1113,12 +1184,12 @@ tinyplot.default = function(
         points(
           x = xx,
           y = yy,
-          col = col[i],
-          bg = bg[i],
-          # type = type, ## rather hardcode "p" to avoid warning message about "pointrange"
+          col = icol,
+          bg = ibg,
+          ## rather hardcode "p" to avoid warning message about "pointrange"
           type = "p",
-          pch = pch[i],
-          lty = lty[i],
+          pch = ipch,
+          lty = ilty,
           cex = cex
         )
       } else if (type %in% c("l", "o", "b", "c", "h", "s", "S", "ribbon")) {
@@ -1127,19 +1198,19 @@ tinyplot.default = function(
         lines(
           x = xx,
           y = yy,
-          col = col[i],
+          col = icol,
           type = type,
-          pch = pch[i],
-          lty = lty[i]
+          pch = ipch,
+          lty = ilty
         )
         if (rtype) type = "ribbon"
       } else if (type == "polygon") {
         polygon(
           x = xx,
           y = yy,
-          border = col[i],
-          col = bg[i],
-          lty = lty[i]
+          border = icol,
+          col = ibg,
+          lty = ilty,
         )
       } else {
         stop("`type` argument not supported.", call. = FALSE)
@@ -1160,7 +1231,7 @@ tinyplot.default = function(
 
 
 #' @rdname tinyplot
-#' @importFrom stats as.formula model.frame
+#' @importFrom stats as.formula model.frame terms
 #' @export
 tinyplot.formula = function(
     x = NULL,
@@ -1217,7 +1288,7 @@ tinyplot.formula = function(
   }
 
   # placeholder for legend title
-  legend.args = list(x = NULL)
+  legend_args = list(x = NULL)
 
   ## set up model frame
   m = match.call(expand.dots = FALSE)
@@ -1292,16 +1363,42 @@ tinyplot.formula = function(
   x = mf[, x_loc]
   by_loc <- x_loc + 1L
   if (NCOL(mf) < by_loc) {
-    by = bylab = NULL  
+    # special catch if by is the same as x or y (normally for continuous legend)
+    by_same_y = by_same_x = FALSE
+    fml_all_vars = all.vars(m$formula, unique = FALSE)
+    if (any(duplicated(fml_all_vars))) {
+      if (isTRUE(no_y)) {
+        by_same_x = TRUE ## i.e., if there is duplication and no y var, assume by must be the same as x
+      } else {
+        fml_lhs_vars = paste(attr(terms(m$formula), "variables")[[2]])
+        fml_rhs_vars = fml_all_vars[!(fml_all_vars %in% fml_lhs_vars)]
+        if (any(duplicated(fml_rhs_vars))) {
+          by_same_x = TRUE
+        } else {
+          by_same_y = TRUE
+        }
+      }
+    }
+    if (isTRUE(by_same_y)) {
+      by = y
+      bylab = names(mf)[y_loc]
+      legend_args[["title"]] = bylab
+    } else if (isTRUE(by_same_x)) {
+      by = x
+      bylab = names(mf)[x_loc]
+      legend_args[["title"]] = bylab
+    } else {
+      by = bylab = NULL
+    }
   } else if (NCOL(mf) == by_loc) {
     by = mf[, by_loc]
     bylab = names(mf)[by_loc]
-    legend.args[["title"]] = bylab
-    if (!inherits(by, "factor")) by = as.factor(by)
+    legend_args[["title"]] = bylab
+    # if (!inherits(by, "factor")) by = as.factor(by)
   } else if (NCOL(mf) > by_loc) {
     by = do.call("interaction", mf[, -c(y_loc, x_loc)])
     bylab = sprintf("interaction(%s)", paste(names(mf)[-c(y_loc, x_loc)], collapse = ", "))
-    legend.args[["title"]] = bylab
+    legend_args[["title"]] = bylab
   }
 
   ## nice axis and legend labels
@@ -1330,7 +1427,7 @@ tinyplot.formula = function(
     frame.plot = frame.plot,
     asp = asp,
     grid = grid,
-    legend.args = legend.args,
+    legend_args = legend_args,
     pch = pch,
     col = col,
     lty = lty,
@@ -1383,9 +1480,9 @@ tinyplot.density = function(
 
   if (inherits(x, "density")) {
     object = x
-    legend.args = list(x = NULL)
+    legend_args = list(x = NULL)
     # Grab by label to pass on legend title to tinyplot.default
-    legend.args[["title"]] = deparse(substitute(by))
+    legend_args[["title"]] = deparse(substitute(by))
   } else {
     ## An internal catch for non-density objects that were forcibly
     ## passed to tinyplot.density (e.g., via a one-side formula)
@@ -1396,7 +1493,7 @@ tinyplot.density = function(
       x = as.numeric(x)
     }
     object = stats::density(x)
-    legend.args = list(...)[["legend.args"]]
+    legend_args = list(...)[["legend_args"]]
   }
 
   by_names = facet_names = NULL
@@ -1493,9 +1590,9 @@ tinyplot.density = function(
     ymin = rep(0, length(y))
     ymax = y
     # set extra legend params to get bordered boxes with fill
-    legend.args[["x.intersp"]] = 1.25
-    legend.args[["lty"]] = 0
-    legend.args[["pt.lwd"]] = 1
+    legend_args[["x.intersp"]] = 1.25
+    legend_args[["lty"]] = 0
+    legend_args[["pt.lwd"]] = 1
   }
 
   ## axes range
@@ -1532,7 +1629,7 @@ tinyplot.density = function(
     frame.plot = frame.plot,
     asp = asp,
     grid = grid,
-    legend.args = legend.args,
+    legend_args = legend_args,
     pch = pch,
     col = col,
     bg = bg,