+++ NEWS.md
@@ -1,7 +1,17 @@
 # News
 
-## 0.0.5.9000 (development version)
+## 0.0.5.9001 (development version)
 
+New features:
+
+- Gradient legends are now supported if a continuous variable is passed to
+`by`. Thanks to @zeileis for detailed feedback and advice around the default
+palette choice (a restricted version of the "viridis" palette), as well as
+StackOverflow user mnel, whose answer
+[here](https://stackoverflow.com/a/13355440) provided the inspiration for the
+final implementation. (#122 @grantmcdermott)
+- Ordered factors now inherit a discrete sequential color palette ("viridis") by
+default. Thanks to @zeileis for the suggestion. (#130 @grantmcdermott)
 - Support user-supplied polygons. (#127 @grantmcdermott)
 
 ## 0.0.5